// @cloud-materials/common/components/CForm/interface.ts
export interface CForm<T extends ObjectType = any, D extends ObjectType = any> extends Form<T> {
  /**
   * @zh 全局数据
   */
  data: D;
  /**
   * @zh 设置全局数据
   */
  setData: (data: Partial<D>) => void;
  /**
   * @zh 标识form名称，对应form标签的name属性
   * @default 随机的唯一uid
   * 默认取formily的form实例id属性
   */
  name: string;
}
// ...
export type ObjectType = Record<string, any>;
// ...
export interface CFormConfirmColumn<T extends ObjectType = any, D extends ObjectType = any>
  extends Omit<ItemProps, 'hidden' | 'label' | 'content'> {
  /**详情页展示字段名 */
  name?: string;
  /**默认根据name取值value，如果需要对value进行改写，可使用formatter */
  formatter?: (value: any, values: T, form?: CForm<T, D>) => ReactNode;
  /**当前字段是否渲染*/
  hidden?: (value: any, values: T, form: CForm<T, D>) => boolean;
  /**label */
  label?: ((value: any, values: T, form?: CForm<T, D>) => ReactNode) | string;
}

/**
 * @title ConfirmSection
 */
export interface ConfirmSection<Column extends any = any, T extends ObjectType = any, F = any> {
  /**当前列title */
  title: string;
  /**当前列是否可编辑，默认不可编辑，需要编辑情况请填入section对应的表单字段*/
  editAnchor?: string;
  /**当前列展示字段集合*/
  fields: Column[];
  /**当前列自定义渲染*/
  customRender?: (values: T, col: ItemProps[], form: F) => ReactNode;
  /**当前列是否渲染*/
  hidden?: (values: T, form: F) => boolean;
  /**当前列配置*/
  columnConfig?: Omit<ColumnConfig<any>, 'title' | 'component' | 'dataIndex' | 'render'>;
}

/**
 * @title ConfirmStep
 */
export interface ConfirmStep<T extends ObjectType = any, D extends ObjectType = any> {
  /**确认步骤title */
  title: string;
  /**确认步骤是否可编辑 */
  edit?: boolean;
  /**确认步骤下展示的列集合 */
  section: ConfirmSection<CFormConfirmColumn<T, D> | string, T, CForm<T, D>>[];
  /**CTable Props */
  CTableProps?: TableConfig<any>;
}

/**
 * @title CFormConfirmDetail
 */
export interface CFormConfirmDetail<T extends ObjectType = any, D extends ObjectType = any> {
  /**详情页信息 */
  detail: ConfirmStep<T, D>[];
  /**详情页协议 */
  protocol?: CAgreementProps & { protocolLabel?: string };
  /**全局设置label宽度 */
  labelWidth?: number;
  /**确认页title，默认为”确认页“ */
  stepTitle?: string;
}
/**
 * @title ConfirmDetail
 */
export interface ConfirmDetail<T extends ObjectType = any, F = any> {
  /**全局设置label宽度 */
  labelWidth?: number;
  /**详情页信息 */
  detail: (Omit<ConfirmStep, 'section'> & { section: ConfirmSection<ItemProps, T, F>[] })[];
  /**可编辑操作返回的步骤回调*/
  onEdit: (step: number, section?: string) => void;
  /**表单values集合*/
  values: T;
  /**表单实例*/
  form: F;
}