// @cloud-materials/common/components/CInfoSection/interface.ts
type Layout = 'horizon' | 'vertical';

/**
 * InfoSection 影响布局的属性
 */
export type CInfoSectionLayout = {
  layout: Layout | Partial<Record<BreakpointEnum, Layout>>;
  colNumber: number | Partial<Record<BreakpointEnum, number>>;
  direction: ('row' | 'column') | Record<BreakpointEnum, 'row' | 'column'>;
};

/**
 * 普通数据类型的 InfoSection 布局
 */
export type CInfoSectionLayoutNormalized = {
  layout: Layout;
  colNumber: number;
  direction: 'row' | 'column';
};

/**
 * @title ItemProps
 */
export interface ItemProps {
  /** 类名 */
  className?: string;
  /** 标签 */
  label?: ReactNode;
  /** 标签对应的内容 支持通用内置组件 */
  content?: DefineBuiltInHybridListType<typeof builtInMap>;
  /** k-v总样式 */
  itemStyle?: CSSProperties;
  /** 标签样式 */
  labelStyle?: CSSProperties;
  /** 内容样式 */
  contentStyle?: CSSProperties;
  /** 控制 label 宽度，栅格布局下不生效 */
  labelWidth?: number | 'normal' | 'small';
  /**
   * 左右或上下结构
   * @defaultValue horizon
   */
  layout?: CInfoSectionLayout['layout'];
  /** 透传内容省略参数，生效需要开启 enableEllipsis */
  cEllipsisProps?: Omit<CEllipsisProps, 'content'>;
  /**
   * 是否开启内容省略
   * @defaultValue false
   */
  enableEllipsis?: boolean;
  /**
   * 当 Content 的内容为空时默认展示的元素
   * @defaultValue -
   */
  defaultEmptyContent?: ReactNode;
  /** 列序号 */
  columnNum?: number;
  /**
   * 动态计算标签宽度
   * @defaultValue true
   */
  auto?: boolean;
  /** 子元素 */
  children?: ReactNode;
  /**
   * 是否隐藏该项，不占位
   * @defaultValue false
   */
  hidden?: boolean;
  /**
   * 是否可见，占位
   * @defaultValue true
   */
  visible?: boolean;
  /** 标签栅格，透传arco props */
  arcoLabelColProps?: ColProps;
  /** 内容栅格，透传arco props */
  arcoContentColProps?: ColProps;
  /** 包裹标签与内容栅格，透传arco props */
  arcoRowProps?: RowProps;
  /** content右侧内容，包含内置组件 */
  extraContent?: DefineBuiltInHybridListType<typeof builtInMap>;
  /** 控制占位数，在 direction 为 row 时生效 */
  span?: number;
}